Separate TileLayers component.
